@@ -18,22 +18,22 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.19.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  bloc: ^6.1.1
-  brasil_fields: ^0.6.0
-  cpf_cnpj_validator: ^1.0.5
+  bloc: ^8.1.4
+  brasil_fields: ^1.14.3
+  cpf_cnpj_validator: ^2.0.0
   cupertino_icons: ^1.0.0
-  dio: ^3.0.10
-  equatable: ^1.2.6
-  flutter_bloc: ^6.1.2
-  flutter_holo_date_picker: ^0.3.1+2
-  flutter_staggered_animations: ^0.1.2
+  dio: ^5.4.3+1
+  equatable: ^2.0.5
+  flutter_bloc: ^8.1.6
+  flutter_holo_date_picker: ^1.1.3
+  flutter_staggered_animations: ^1.1.1
   pedantic: ^1.9.0
-  selectable_circle: ^0.1.6+4
+  #selectable_circle: ^0.1.6+4
   
 
 dev_dependencies:
