@@ -9,7 +9,7 @@ import '../app_import.dart';
 // PRIMEIRA PAGINA CARREGA COM A LOGO DEPOIS NAVEGA PARA PAGINA COM A LISTA DE REGISTROS
 class SplashPage extends StatefulWidget {
   final String title;
-  const SplashPage({Key key, this.title = "Splash"}) : super(key: key);
+  const SplashPage({Key? key, this.title = "Splash"}) : super(key: key);
 
   @override
   _SplashPageState createState() => _SplashPageState();
