@@ -1,15 +1,14 @@
 import 'package:bloc/bloc.dart';
 
 class SimpleBlocObserver extends BlocObserver {
-  @override
-  void onError(Cubit cubit, Object error, StackTrace stackTrace) {
-    print(error);
-    super.onError(cubit, error, stackTrace);
-  }
+  // @override
+  // void onError(Cubit cubit, Object error, StackTrace stackTrace) {
+  //   print(error);
+  //   super.onError(cubit, error, stackTrace);
+  // }
 
-  @override
-  void onChange(Cubit cubit, Change change) {
-    print(change);
-    super.onChange(cubit, change);
-  }
+  // void onChange(Cubit cubit, Change change) {
+  //   print(change);
+  //   super.onChange(cubit, change);
+  // }
 }
